added interfaces

+++ Showroom.Domain/Entities/Generation.cs
@@ -8,7 +8,7 @@
         }
 
         public int IdGeneration { get; set; }
-        public int Year { get; set; }
+        public Int16 Year { get; set; }
         public int Produced { get; set; }
 
         public virtual ICollection<Model> Models { get; set; }
